cleaning up code

+++ main/forms.py
@@ -6,4 +6,4 @@ class PostNewCommand(forms.Form):
     par = forms.CharField(label="par", max_length=1)
     int = forms.CharField(label="int", max_length=1)
     mod = forms.IntegerField(min_value=0, max_value=255)
-    val = forms.IntegerField()
+    val = forms.FloatField()
